pr comments, moved things around including new bc utils module that will store actual utilities and avoid recursive import from parse table
